Adjust spell check tool configuration to avoid false positive
The repository's CI system includes a check for commonly misspelled words.

The latest version of the codespell tool dependency used for this spell check is always used. This means that the
results of the spell check can change as new versions of the tool are released. The recent codespell release introduced
a false positive result. This is corrected by adding the word to the ignore list in the codespell configuration file.