Allow the fixed fonts to be scaled up.
On larger displays, the two built in fonts are almost invisible.
One possible solution is to supply some larger fonts.

Another approach is to do, like several TFT display drivers do and allow you to scale the system fixed fonts up to a larger size.  Decided this was the easiest approach, and did a quick implementation of it.  Like the Adafruit displays, I allowed you to set different scale factors for X and Y.

I added a scaledBitmap function which is a modified version of the current bitmap function.  I also changed all of the text functions to use it.  I left the unscalled version of the API, although I have it simply call the scaled version with scales 1 and 1, as since these methods are virtual, potentially some other subclasses might use and/or implement their own version.

Updated the textFontWidth and textFontHeight to multiply  the height * the scale as
the one example sketch here needs it.

I also updated the sketch to scale the canvas to take the font size into account.
This appears to work with the changes and

I have not used or tested the scroll code.  Its sizing should be updated teh same way
the test sketch was with the changes with the textFontWidth and textFontHeight changes

Update api.md

Forgot to update the AsciiDraw example sketch

Code Review requested changes

_textsize_x -> _textSizeX  (ditto for y)
setTextSize() -> textSize()

scaledBitmap -> bitmap with the extra parameters with defaults