Correct scroll length for left and up directions
Previously, the calculation for the distance of leftward and upward scrolling resulted in the text bitmap not being
scrolled completely off the display.

For leftward scrolling, this would be noticeable in the case where the bitmap of the last character in the string had
populated pixels on the rightmost column. For upward scrolling, this would be noticeable in the case where the bitmap of
any character in the string had populated pixels on the bottom row.